Issue788TasksForExtractors

+++ openrem/remapp/tools/background.py
@@ -35,12 +35,14 @@ import time
 import signal
 import sys
 import uuid
+import random
 from multiprocessing import Process
-import datetime
 
 import django
 from django import db
 from django.db.models import Q
+from django.db import transaction
+from django.utils import timezone
 
 # Setup django. This is required on windows, because process is created via spawn and
 # django will not be initialized anymore then (On Linux this will only be executed once)
@@ -51,10 +53,10 @@ if projectpath not in sys.path:
 os.environ["DJANGO_SETTINGS_MODULE"] = "openremproject.settings"
 django.setup()
 
-from remapp.models import BackgroundTask
+from remapp.models import BackgroundTask, DicomQuery
 
 
-def run_as_task(func, task_type, taskuuid, *args, **kwargs):
+def run_as_task(func, task_type, num_proc, num_of_task_type, taskuuid, *args, **kwargs):
     """
     Runs func as a task. (Which means it runs normally, but a BackgroundTask
     object is created and hence the execution as well as occurred errors are
@@ -62,11 +64,25 @@ def run_as_task(func, task_type, taskuuid, *args, **kwargs):
 
     As a note: This is used as a helper for run_in_background. However, in
     principle it could also be used to run any function in sequential
-    for which we would like to document that it was executed.
+    for which we would like to document that it was executed. (Has some not so nice
+    parts, especially that user can terminate a bunch of sequentially running tasks)
+
+    Note that waiting here if it is ok to start is actually quite ugly, since running a lot
+    of processes with conditions will use a lot of RAM without any use. This however is the
+    simplest possible fix for making the run_in_background_with_limits non-blocking, which
+    is a requirement for the docker import. A nice update would be to either have 1 process
+    managing the execution of the other processes, or at least do this based on signalling
+    from the exiting processes instead of polling.
+
+    This can sometimes lead to errors with sqlite, see (tasks keep hanging then)
+    https://stackoverflow.com/questions/28958580/django-sqlite-database-is-locked
 
     :param func: The function to run
     :param task_type: A string documenting what kind of task this is
-    :param taskuuid: An uuid which will be used as uuid of the BackgroundTask object. If None will generate one itself
+    :param num_proc: The maximum number of processes that should be executing
+    :param num_of_task_type: The maximum number of processes for multiple tasks type which are allowed to
+        run at the same time
+    :param taskuuid: An uuid which will be used as uuid of the BackgroundTask object. If None, will generate one itself
     :args: Args to func
     :kwargs: Args to func
     :return: The created BackgroundTask object
@@ -78,9 +94,35 @@ def run_as_task(func, task_type, taskuuid, *args, **kwargs):
         uuid=taskuuid,
         proc_id=os.getpid(),
         task_type=task_type,
-        started_at=datetime.datetime.now(),
+        started_at=None,
     )
     b.save()
+
+    if num_proc > 0 or len(num_of_task_type) > 0:
+        while True:
+            with transaction.atomic():
+                qs = BackgroundTask.objects.filter(
+                    complete__exact=False
+                )
+                next_wait_time = qs.count()
+                a = qs.filter(started_at__isnull=False).count()
+                if num_proc == 0 or a < num_proc:
+                    ok = True
+                    for k, v in num_of_task_type.items():
+                        a = qs.filter(
+                            Q(task_type__exact=k) & Q(started_at__isnull=False)
+                        ).count()
+                        ok = ok and a < v
+                        if not ok:
+                            break
+                    if ok:
+                        b.started_at = timezone.now()
+                        b.save()
+                        break
+            time.sleep(0.2 + 0.4 * next_wait_time * random.random())
+    else:
+        b.started_at = timezone.now()
+        b.save()
 
     try:
         func(*args, **kwargs)
@@ -100,7 +142,9 @@ def run_as_task(func, task_type, taskuuid, *args, **kwargs):
     return b
 
 
-def run_in_background(func, task_type, *args, **kwargs):
+def run_in_background_with_limits(
+    func, task_type, num_proc, num_of_task_type, *args, **kwargs
+):
     """
     Runs fun as background Process.
 
@@ -112,6 +156,7 @@ def run_in_background(func, task_type, *args, **kwargs):
     complete flag will be set to True.
     This function cannot be used with Django Tests, unless they use TransactionTestCase
     instead of TestCase (which is far slower, so use with caution).
+    num_proc and num_of_task_type can be used to give conditions on the start.
 
     :param func: The function to run. Note that you should set the status of the task yourself
         and mark as completed when exiting yourself e.g. via sys.exit(). Assuming the function
@@ -119,17 +164,25 @@ def run_in_background(func, task_type, *args, **kwargs):
         the BackgroundTask object will be set correctly.
     :param task_type: One of the strings declared in BackgroundTask.task_type. Indicates which
         kind of background process this is supposed to be. (E.g. move, query, ...)
+    :param num_proc: Will wait with execution until there are less than num_proc active tasks.
+        If num_proc == 0 will run directly
+    :param num_of_task_type: A dictionary from str to int, where key should be some used
+        task_type. Will wait until there are less active tasks of task_type than
+        specified in the value of the dict for that key.
     :param args: Positional arguments. Passed to func.
     :param kwargs:  Keywords arguments. Passed to func.
     :returns: The BackgroundTask object.
     """
+    taskuuid = str(uuid.uuid4())
+
     # On linux connection gets copied which leads to problems.
     # Close them so a new one is created for each process
     db.connections.close_all()
 
-    taskuuid = str(uuid.uuid4())
     p = Process(
-        target=run_as_task, args=(func, task_type, taskuuid, *args), kwargs=kwargs
+        target=run_as_task,
+        args=(func, task_type, num_proc, num_of_task_type, taskuuid, *args),
+        kwargs=kwargs,
     )
 
     p.start()
@@ -147,35 +200,59 @@ def run_in_background(func, task_type, *args, **kwargs):
     return _get_task_via_uuid(taskuuid)
 
 
+def run_in_background(func, task_type, *args, **kwargs):
+    """
+    Syntactic sugar around run_in_background_with_limits.
+
+    Arguments correspond to run_in_background_with_limits. Will always run the
+    passed function as fast as possible. This function should always
+    be used for functions triggered from the webinterface.
+    """
+    return run_in_background_with_limits(func, task_type, 0, {}, *args, **kwargs)
+
+
 def terminate_background(task: BackgroundTask):
     """
     Terminate a background task by force. Sets complete=True on the task object.
     """
-    try:
-        if os.name == "nt":
-            # On windows this signal is not implemented. The api will just use TerminateProcess instead.
-            os.kill(task.proc_id, signal.SIGTERM)
-        else:
-            os.kill(task.proc_id, signal.SIGTERM)
-            # Wait until the process has returned (potentially it already has when we call wait)
-            # On  Windows the equivalent does not work, but seems to be blocking there anyway
-            os.waitpid(task.proc_id, 0)
-
-    except (ProcessLookupError, OSError):
-        pass
     task.completed_successfully = False
     task.complete = True
     task.error = "Forcefully aborted"
     task.save()
+
+    if task.proc_id > 0:
+        try:
+            if os.name == "nt":
+                # On windows this signal is not implemented. The api will just use TerminateProcess instead.
+                os.kill(task.proc_id, signal.SIGTERM)
+            else:
+                os.kill(task.proc_id, signal.SIGTERM)
+                # Wait until the process has returned (potentially it already has when we call wait)
+                # On  Windows the equivalent does not work, but seems to be blocking there anyway
+                os.waitpid(task.proc_id, 0)
+
+        except (ProcessLookupError, OSError):
+            pass
+
+
+def wait_task(task: BackgroundTask):
+    """
+    Wait until the task has completed
+    """
+    while True:
+        task.refresh_from_db()
+        if task.complete:
+            return
+        time.sleep(0.3)
 
 
 def _get_task_via_uuid(task_uuid):
     return BackgroundTask.objects.filter(uuid__exact=task_uuid).first()
 
 
-def _get_task_via_pid(proc_id):
+def _get_task_via_pid(process_id):
     return BackgroundTask.objects.filter(
-        Q(proc_id__exact=proc_id) & Q(complete__exact=False)
+        Q(proc_id__exact=process_id) & Q(complete__exact=False)
     ).first()
 
 
@@ -224,3 +301,29 @@ def record_task_error_exit(error_msg):
         b.completed_successfully = False
         b.error = error_msg
         b.save()
+
+
+def record_task_related_query(study_instance_uid):
+    """
+    Tries to find the related DicomQRRspStudy object
+    given a study instance uid and if this is running in
+    a task will record it to the query object. This
+    is used to later find the import tasks that were run as
+    part of a query.
+    Since this actually just takes the latest query if the user
+    runs imports manually via script it may in principle wrongly
+    associate.
+    """
+    b = get_current_task()
+    if b is not None:
+        if DicomQuery.objects.exists():
+            queries = DicomQuery.objects.filter(
+                started_at__isnull=False
+            ).latest(
+                "started_at"
+            ).dicomqrrspstudy_set.filter(
+                study_instance_uid=study_instance_uid
+            ).all()
+
+            for query in queries:
+                query.related_imports.add(b)
